refactor test object namespaces
