+++ DragonFruit.Common.Data.Tests/Requests/RequestTests.cs
@@ -3,7 +3,6 @@
 
 using DragonFruit.Common.Data.Basic;
 using DragonFruit.Common.Data.Extensions;
-using DragonFruit.Common.Data.Tests.Requests.Objects;
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
